// src/services/controlPlaneApi/gen/models/application.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Project API
 * Project description API
 *
 * The version of the OpenAPI document: undefined
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { Dependency } from './dependency';
// May contain unused imports in some cases
// @ts-ignore
import { Gateways } from './gateways';
// May contain unused imports in some cases
// @ts-ignore
import { Instance } from './instance';
// May contain unused imports in some cases
// @ts-ignore
import { Module } from './module';
// May contain unused imports in some cases
// @ts-ignore
import { Resource } from './resource';
// May contain unused imports in some cases
// @ts-ignore
import { Secrets } from './secrets';

/**
 * 
 * @export
 * @interface Application
 */
export interface Application {
    /**
     * 
     * @type {{ [key: string]: Resource; }}
     * @memberof Application
     */
    'resources'?: { [key: string]: Resource; };
    /**
     * 
     * @type {{ [key: string]: Module; }}
     * @memberof Application
     */
    'modules'?: { [key: string]: Module; };
    /**
     * 
     * @type {Array<Dependency>}
     * @memberof Application
     */
    'dependencies'?: Array<Dependency>;
    /**
     * 
     * @type {Gateways}
     * @memberof Application
     */
    'gateways'?: Gateways;
    /**
     * 
     * @type {Instance}
     * @memberof Application
     */
    'instance'?: Instance;
    /**
     * 
     * @type {Secrets}
     * @memberof Application
     */
    'secrets'?: Secrets;
}