// src/services/controlPlaneApi/gen/models/compute-cluster.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Project API
 * Project description API
 *
 * The version of the OpenAPI document: undefined
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ComputeCluster
 */
export interface ComputeCluster {
    /**
     * 
     * @type {string}
     * @memberof ComputeCluster
     */
    'type'?: string;
    /**
     * 
     * @type {{ [key: string]: object; }}
     * @memberof ComputeCluster
     */
    'configuration'?: { [key: string]: object; };
}