// src/services/controlPlaneApi/gen/models/streaming-cluster.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Project API
 * Project description API
 *
 * The version of the OpenAPI document: undefined
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface StreamingCluster
 */
export interface StreamingCluster {
    /**
     * 
     * @type {string}
     * @memberof StreamingCluster
     */
    'type'?: string;
    /**
     * 
     * @type {{ [key: string]: object; }}
     * @memberof StreamingCluster
     */
    'configuration'?: { [key: string]: object; };
}