// src/services/controlPlaneApi/gen/models/topic-definition.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Project API
 * Project description API
 *
 * The version of the OpenAPI document: undefined
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { SchemaDefinition } from './schema-definition';

/**
 * 
 * @export
 * @interface TopicDefinition
 */
export interface TopicDefinition {
    /**
     * 
     * @type {string}
     * @memberof TopicDefinition
     */
    'name'?: string;
    /**
     * 
     * @type {{ [key: string]: object; }}
     * @memberof TopicDefinition
     */
    'config'?: { [key: string]: object; };
    /**
     * 
     * @type {{ [key: string]: object; }}
     * @memberof TopicDefinition
     */
    'options'?: { [key: string]: object; };
    /**
     * 
     * @type {SchemaDefinition}
     * @memberof TopicDefinition
     */
    'keySchema'?: SchemaDefinition;
    /**
     * 
     * @type {SchemaDefinition}
     * @memberof TopicDefinition
     */
    'valueSchema'?: SchemaDefinition;
    /**
     * 
     * @type {number}
     * @memberof TopicDefinition
     */
    'partitions'?: number;
    /**
     * 
     * @type {boolean}
     * @memberof TopicDefinition
     */
    'implicit'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof TopicDefinition
     */
    'creation-mode'?: string;
}